Delete bid files
Moved things around more, renamed the bid function to bidUntilContract
