+++ src/cardUtils.c
@@ -71,6 +71,41 @@ Bool setCanPlay(Card cardArray[], int nbOfCards, Color conditionalColor, Color t
     return conditionMet;
 }
 
+void findValidCardsInHand(Card cardsInHand[], int nbOfCardsInHand, Card trickCards[], int nbOfTrickCards, Color trump) {
+    Bool canFollow;
+    Card bestCard;
+    int bestTrumpStrength = 0;
+
+    if (nbOfTrickCards == 0) {                                                                                      //The first player of a trick
+        setCanPlay(cardsInHand, nbOfCardsInHand, NULL_COLOR, trump, 0, TRUE);                                       //can play any card.
+    }
+    else {
+        if (trump == ALLTRUMP) {
+            trump = trickCards[0].color;                                                                            //This line is what makes this function work when the trump is ALLTRUMP
+        }
+        bestCard = trickCards[getStrongestCard(trickCards, nbOfTrickCards, trump, NULL_COLOR)];                     //Find the best card on the table
+        if (bestCard.color == trump) {                                                                              //If it's a trump,
+            bestTrumpStrength = getCardStrength(bestCard, trump, NULL_COLOR);                                       //update bestTrumpStrength
+        }
+        setCanPlay(cardsInHand, nbOfCardsInHand, NULL_COLOR, trump, 0, FALSE);                                      //Each card is initialised to canPlay = FALSE
+        canFollow = setCanPlay(cardsInHand, nbOfCardsInHand, trickCards[0].color, trump, bestTrumpStrength, TRUE);  //A player must follow in the right color. If that color is a trump, the player has to play a stronger card. 
+        if ((trickCards[0].color == trump) && (canFollow == FALSE)) {                                               //If the trick was started with a trump AND playing a higher trump than the current best one is impossible,
+            canFollow = setCanPlay(cardsInHand, nbOfCardsInHand, trump, trump, 0, TRUE);                            //the player has to follow with a lower trump.
+        }
+        if (canFollow == FALSE) {                                                                                   //If playing in the right color is impossible:
+            if (getStrongestCard(trickCards, nbOfTrickCards, trump, trickCards[0].color) == (nbOfTrickCards - 2)) { //If the player's partner is the current trick winner,
+                setCanPlay(cardsInHand, nbOfCardsInHand, NULL_COLOR, trump, 0, TRUE);                               //any card can be played.
+            }
+            else {
+                canFollow = setCanPlay(cardsInHand, nbOfCardsInHand, trump, trump, bestTrumpStrength, TRUE);        //A player must play a stronger trump card than the current best one if its partner isn't winning.
+                if (canFollow == FALSE) {                                                                           //If all of the above are impossible,
+                    setCanPlay(cardsInHand, nbOfCardsInHand, NULL_COLOR, trump, 0, TRUE);                           //the player can play any card.
+                }
+            }
+        }
+    }    
+}
+
 void sortCards(Card cardArray[], int nbToSort, Color trump, Color roundColor) { //Bubble sort algorithm
     Card previousCard;
     int newNbToSort;
