+++ src/play.h
@@ -2,14 +2,20 @@
 #define _PLAY_H_
 #include "core.h"
 
-/* Sets the canPlay property of each card in an Array to find which ones can be played depending on the cards currently on the table
-*   @param cardsInHand[]: the cards in this array will be evaluated to find the valid ones to play according to the coinche rules
-*   @param nbOfCardsInHand: the number of cards this function should look at to see which ones are valid
-*   @param trickCards[]: the cards already played by the previous players. Only the first N cards matter, where N is nbOfTrickCards
-*   @param nbOfTrickCards: the number of cards played by the previous players. Can be 0
-*   @param trump: the current trump
+/* Play a bid attempt: either a contract is made or everyone passed
+*   @param players[]: array of 4 players
+*   @param startingPlayer: position of the starting player
+*   @param *contract: pointer to the contract being debated. Will be edited someone decides to make a contract
+*   @return everyonePassed: FALSE if a contract was made, TRUE otherwise
 */
-void findValidCardsInHand(Card cardsInHand[], int nbOfCardsInHand, Card trickCards[], int nbOfTrickCards, Color trump);
+Bool bidAttempt(Player players[], Position startingPlayer, Contract *contract);
+
+/* Deal cards and play bid attempts until a contract is made
+*   @param players[]: array of 4 players
+*   @param startingPlayer: position of the starting player
+*   @return contract: the contract that was made
+*/
+Contract bidUntilContract(Player players[], Position startingPlayer);
 
 /* Plays a single trick : each player plays a card, then the player with the strongest card wins the trick and gets points
 *   @param players[]: array of 4 players
